Move MqttData to models.py
