+++ homeassistant/components/mqtt/diagnostics.py
@@ -16,7 +16,8 @@ from homeassistant.core import HomeAssistant, callback, split_entity_id
 from homeassistant.helpers import device_registry as dr, entity_registry as er
 from homeassistant.helpers.device_registry import DeviceEntry
 
-from . import debug_info, get_mqtt_data, is_connected
+from . import debug_info, is_connected
+from .models import get_mqtt_data
 
 REDACT_CONFIG = {CONF_PASSWORD, CONF_USERNAME}
 REDACT_STATE_DEVICE_TRACKER = {ATTR_LATITUDE, ATTR_LONGITUDE}
